federation: tests for findTasks and some refactoring
extract common test logic to be a bit more reusable
