+++ titus-server-federation/src/test/java/com/netflix/titus/federation/service/AggregatingAutoScalingServiceTest.java
@@ -65,7 +65,7 @@ public class AggregatingAutoScalingServiceTest extends AggregatingAutoScalingTes
         ScalingPolicyResult policyOneResult = ScalingPolicyResult.newBuilder().setId(policy1).build();
 
         CellWithPolicies cellOneService = new CellWithPolicies(Collections.singletonList(policyOneResult));
-        FailingCell badCell = new FailingCell();
+        CellWithFailingAutoscalingService badCell = new CellWithFailingAutoscalingService();
 
         cellOne.getServiceRegistry().addService(cellOneService);
         cellTwo.getServiceRegistry().addService(badCell);
@@ -115,8 +115,8 @@ public class AggregatingAutoScalingServiceTest extends AggregatingAutoScalingTes
         CellWithPolicies cellOneService = new CellWithPolicies(Collections.singletonList(policyOneResult));
         CellWithPolicies cellTwoService = new CellWithPolicies(Collections.singletonList(policyTwoResult));
 
-        CellWithJobs cellOneJobsService = new CellWithJobs(Collections.singletonList(JOB_1));
-        CellWithJobs cellTwoJobsService = new CellWithJobs(Collections.singletonList(JOB_2));
+        CellWithJobIds cellOneJobsService = new CellWithJobIds(Collections.singletonList(JOB_1));
+        CellWithJobIds cellTwoJobsService = new CellWithJobIds(Collections.singletonList(JOB_2));
         cellOne.getServiceRegistry().addService(cellOneService);
         cellOne.getServiceRegistry().addService(cellOneJobsService);
         cellTwo.getServiceRegistry().addService(cellTwoService);
@@ -174,8 +174,8 @@ public class AggregatingAutoScalingServiceTest extends AggregatingAutoScalingTes
 
         CellWithPolicies cellOneService = new CellWithPolicies(Collections.singletonList(policyOneResult));
         CellWithPolicies cellTwoService = new CellWithPolicies(Collections.singletonList(policyTwoResult));
-        CellWithJobs cellOneJobsService = new CellWithJobs(Collections.singletonList(JOB_1));
-        CellWithJobs cellTwoJobsService = new CellWithJobs(Collections.singletonList(JOB_2));
+        CellWithJobIds cellOneJobsService = new CellWithJobIds(Collections.singletonList(JOB_1));
+        CellWithJobIds cellTwoJobsService = new CellWithJobIds(Collections.singletonList(JOB_2));
 
         cellOne.getServiceRegistry().addService(cellOneService);
         cellTwo.getServiceRegistry().addService(cellOneJobsService);
