+++ titus-server-federation/src/test/java/com/netflix/titus/federation/service/AggregatingAutoScalingTestBase.java
@@ -15,35 +15,18 @@
  */
 package com.netflix.titus.federation.service;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.UUID;
-import java.util.stream.Collectors;
 
-import com.google.protobuf.Empty;
 import com.netflix.titus.api.federation.model.Cell;
 import com.netflix.titus.common.grpc.AnonymousSessionContext;
 import com.netflix.titus.federation.startup.GrpcConfiguration;
-import com.netflix.titus.grpc.protogen.DeletePolicyRequest;
-import com.netflix.titus.grpc.protogen.GetPolicyResult;
-import com.netflix.titus.grpc.protogen.Job;
-import com.netflix.titus.grpc.protogen.JobId;
-import com.netflix.titus.grpc.protogen.PutPolicyRequest;
-import com.netflix.titus.grpc.protogen.ScalingPolicyID;
-import com.netflix.titus.grpc.protogen.ScalingPolicyResult;
-import com.netflix.titus.grpc.protogen.UpdatePolicyRequest;
 import io.grpc.ManagedChannel;
-import io.grpc.Status;
-import io.grpc.stub.StreamObserver;
 import io.grpc.testing.GrpcServerRule;
 import org.junit.Before;
 import org.junit.Rule;
 
-import static com.netflix.titus.grpc.protogen.AutoScalingServiceGrpc.AutoScalingServiceImplBase;
-import static com.netflix.titus.grpc.protogen.JobManagementServiceGrpc.JobManagementServiceImplBase;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -77,102 +60,6 @@ public class AggregatingAutoScalingTestBase {
         GrpcConfiguration grpcConfiguration = mock(GrpcConfiguration.class);
         when(grpcConfiguration.getRequestTimeoutMs()).thenReturn(1000L);
         service = new AggregatingAutoScalingService(connector, new AnonymousSessionContext(), grpcConfiguration);
-    }
-
-    static class CellWithJobs extends JobManagementServiceImplBase {
-        private List<String> jobIds;
-
-        CellWithJobs(List<String> jobIds) {
-            this.jobIds = jobIds;
-        }
-
-        @Override
-        public void findJob(JobId request, StreamObserver<Job> responseObserver) {
-            if (jobIds.contains(request.getId())) {
-                responseObserver.onNext(Job.newBuilder().setId(request.getId()).build());
-                responseObserver.onCompleted();
-            } else {
-                responseObserver.onError(Status.NOT_FOUND.asException());
-            }
-        }
-    }
-
-    static class CellWithPolicies extends AutoScalingServiceImplBase {
-        private final Map<String, ScalingPolicyResult> policyMap;
-
-        CellWithPolicies(List<ScalingPolicyResult> policyResults) {
-            policyMap = policyResults.stream().collect(Collectors.toMap(p -> p.getId().getId(), p -> p));
-        }
-
-        @Override
-        public void getAllScalingPolicies(Empty request, StreamObserver<GetPolicyResult> responseObserver) {
-            sendScalingPolicyResults(new ArrayList<>(policyMap.values()), responseObserver);
-        }
-
-        @Override
-        public void getJobScalingPolicies(JobId request, StreamObserver<GetPolicyResult> responseObserver) {
-            List<ScalingPolicyResult> scalingPolicyResults = policyMap.values().stream()
-                    .filter(p -> p.getJobId().equals(request.getId()))
-                    .collect(Collectors.toList());
-            sendScalingPolicyResults(scalingPolicyResults, responseObserver);
-        }
-
-        @Override
-        public void getScalingPolicy(ScalingPolicyID request, StreamObserver<GetPolicyResult> responseObserver) {
-            List<ScalingPolicyResult> result = policyMap.values().stream()
-                    .filter(p -> p.getId().getId().equals(request.getId())).collect(Collectors.toList());
-            sendScalingPolicyResults(result, responseObserver);
-        }
-
-        @Override
-        public void setAutoScalingPolicy(PutPolicyRequest request, StreamObserver<ScalingPolicyID> responseObserver) {
-            ScalingPolicyID newPolicyId = ScalingPolicyID.newBuilder().setId(UUID.randomUUID().toString()).build();
-            ScalingPolicyResult newPolicyResult = ScalingPolicyResult.newBuilder().setId(newPolicyId).setJobId(request.getJobId()).build();
-
-            policyMap.put(newPolicyId.getId(), newPolicyResult);
-            responseObserver.onNext(newPolicyId);
-            responseObserver.onCompleted();
-        }
-
-        @Override
-        public void updateAutoScalingPolicy(UpdatePolicyRequest request, StreamObserver<Empty> responseObserver) {
-            if (policyMap.containsKey(request.getPolicyId().getId())) {
-                ScalingPolicyResult currentPolicy = policyMap.get(request.getPolicyId().getId());
-                ScalingPolicyResult updatePolicy = ScalingPolicyResult.newBuilder().setId(currentPolicy.getId()).setJobId(currentPolicy.getJobId())
-                        .setScalingPolicy(request.getScalingPolicy()).build();
-                policyMap.put(currentPolicy.getId().getId(), updatePolicy);
-                responseObserver.onNext(Empty.getDefaultInstance());
-                responseObserver.onCompleted();
-            } else {
-                responseObserver.onError(new IllegalArgumentException(request.getPolicyId().getId() + " is Invalid"));
-            }
-        }
-
-        @Override
-        public void deleteAutoScalingPolicy(DeletePolicyRequest request, StreamObserver<Empty> responseObserver) {
-            if (policyMap.containsKey(request.getId().getId())) {
-                policyMap.remove(request.getId().getId());
-                responseObserver.onNext(Empty.getDefaultInstance());
-                responseObserver.onCompleted();
-            } else {
-                responseObserver.onError(new IllegalArgumentException(request.getId().getId() + " is Invalid"));
-            }
-        }
-
-        private void sendScalingPolicyResults(List<ScalingPolicyResult> results,
-                                              StreamObserver<GetPolicyResult> responseObserver) {
-            GetPolicyResult.Builder getPolicyResultBuilder = GetPolicyResult.newBuilder();
-            results.forEach(getPolicyResultBuilder::addItems);
-            responseObserver.onNext(getPolicyResultBuilder.build());
-            responseObserver.onCompleted();
-        }
-    }
-
-    static class FailingCell extends AutoScalingServiceImplBase {
-        @Override
-        public void getAllScalingPolicies(Empty request, StreamObserver<GetPolicyResult> responseObserver) {
-            responseObserver.onError(Status.UNAVAILABLE.asRuntimeException());
-        }
     }
 
 }
